# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: Yuva API
  version: 0.0.3
  summary: Customer messaging for many products in one inbox.
  license:
    name: MIT
    identifier: MIT
  description: |
    Yuva answers every user of every product you run: e-mail, live chat and in-app
    conversations in one inbox.

    ## Surfaces

    - `/v1` is for members (panel session) and workspace API keys.
    - `/client/v1` is for the web widget and the mobile SDKs.
    - `/healthz` and `/readyz` are for load balancers and orchestrators.

    ## Errors

    Errors are `application/problem+json` documents (RFC 9457) with a stable `code`.

    ## Authentication

    - Members sign in with a one-time code sent by e-mail or with a passkey. Both set the
      `yuva_session` cookie (HttpOnly, SameSite=Lax).
    - Host backends, scripts and bots send a workspace API key as `Authorization: Bearer <key>`.
      A key acts on its own workspace only, and only with its scopes (see `api-keys`).
    - Contacts (widget, SDKs) send a contact session token as `Authorization: Bearer <token>` on
      `/client/v1`; see the `client` tag.

    ## Idempotency

    Every authenticated `POST` on `/v1` and `/client/v1` accepts an `Idempotency-Key` header (1 to
    255 printable ASCII characters). The key is remembered for 24 hours per caller (API key,
    member or contact) with the method, path, a SHA-256 of the body and the response. The same key
...
    rather than a workspace (`/v1/me/...`, `/v1/auth/...`), ignore the header.

...

...
servers:
...
security:
  - session: []
  - apiKey: []
...
paths:
...
  /v1/inboxes/{inboxId}:
    get:
      operationId: getInbox
      summary: Get an inbox
      description: |
        An inbox the caller cannot see answers `404`.

        Scope: `inboxes:read`.
      tags: [inboxes]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/InboxId"
      responses:
        "200":
          description: The inbox.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateInbox
      summary: Update an inbox
      description: |
        Owners, admins and API keys.

        Scope: `inboxes:manage`.
      tags: [inboxes]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/InboxId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboxUpdate"
      responses:
        "200":
          description: The updated inbox.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inbox"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteInbox
      summary: Delete an inbox
      description: |
        Owners, admins and API keys. Deletes its channels, conversations,
        messages and attachments.

        Scope: `inboxes:manage`.
      tags: [inboxes]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/InboxId"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
  /v1/realtime:
    get:
      operationId: realtime
      summary: Live events (WebSocket)
      description: |
        Upgrades to a WebSocket that streams the workspace's events as they are committed. Members
        authenticate with the session cookie; the `Origin` header must then be the server's public
        origin (or one of the configured WebAuthn origins). API keys send
        `Authorization: Bearer <key>`. Browsers cannot set headers on a WebSocket, so the workspace
        can also be named with the `workspace_id` query parameter instead of `Yuva-Workspace`.

        Every server message is one JSON text frame matching `RealtimeMessage`. Events carry an
        increasing `id`; within a workspace ids are in commit order. Owners, admins and API keys
        receive every event of the workspace. Agents, and API keys limited to inboxes, receive the
        events of the inboxes they can access and contact events; agents also get
        `inbox_access.changed` events about themselves. Access is re-read when such an event
        arrives and every 30 seconds. An API key needs `conversations:read`. `inbox.created` reaches owners,
        admins and API keys (an agent learns of a new inbox from `inbox_access.changed`);
        `inbox.deleted` reaches everyone who could see the inbox. `conversation.read` reaches only
        the connections of the member who read. `typing` (a contact or another member typing) has no
        `id`, is not stored and is not replayed.

        A member session's open connection also makes the member available to contacts of `live`
        inboxes they can access (see `availability` on `/v1/me`).

        Resuming: after the replay (if any) the server sends `ready` with `last_event_id`, the
        stream position at that moment. Remember the larger of that value and the `id` of every
        event received, and reconnect with `?last_event_id=<it>`. The server then replays the
        events after it, without gaps or duplicates, before the live ones. Events are kept for 24
        hours; when `last_event_id` is no longer known the server sends `resync_required` (reload
        everything over HTTP) followed by `ready` and continues live. Without `last_event_id` there
        is no replay.

        The server pings every 30 seconds. A member's panel reports the conversation it shows with
        a text frame `{"type": "viewing", "conversation_id": "<id>"}`, and
        `{"type": "viewing", "conversation_id": null}` when it shows none or the page is hidden;
        send it again after every reconnect. While a connection of the member views a
        conversation, the member gets no push or e-mail notification about it. Other client
        frames are ignored. Close codes: 1008 (`unauthenticated`, `forbidden`) when the session
        or membership ends, 1013 (`slow_consumer`) when the client does not read fast enough, 1012
        (`restart`) when the server restarts or loses its event feed. Reconnect with
        `last_event_id` after any of them except 1008. While the server is reconnecting its own event
        feed the upgrade is refused with 503 `realtime_unavailable`; retry with backoff.
      tags: [realtime]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - name: workspace_id
          in: query
          description: The workspace to act on, for clients that cannot send `Yuva-Workspace`.
          schema:
            type: string
            format: uuid
        - name: last_event_id
          in: query
          description: Replay the events after this one before streaming live ones.
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        "101":
          description: |
            Switching to the WebSocket protocol. Each text frame is one `RealtimeMessage`; the
            content below describes a frame, not an HTTP body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealtimeMessage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: yuva_session
      description: |
        Member session set by a successful sign-in. Requests other than `GET` and `HEAD` with this
        cookie must send an `Origin` of the server's public origin (or a WebAuthn origin), or
        `Sec-Fetch-Site: same-origin` without `Origin` (`403 origin_not_allowed`), and a body as
        `application/json` or `multipart/form-data` (`415 unsupported_media_type`).
    apiKey:
      type: http
      scheme: bearer
      description: |
        A workspace API key, `yuva_<prefix>_<secret>`. Each operation a key may call needs one
        scope (`ApiKeyScope`); a call without it is refused with `403 insufficient_scope` naming the
        missing scope in `scope`. Operations that list no scope need a member session
        (`403 member_session_required`). An expired key is refused with `401 api_key_expired`.
    contactSession:
      type: http
      scheme: bearer
      description: A contact session token from `POST /client/v1/session`, `yuva_cs_<secret>`.
  parameters:
    WorkspaceHeader:
      name: Yuva-Workspace
      in: header
      required: false
      description: The workspace to act on; see "Workspace selection".
      schema:
        type: string
        format: uuid
...
    InboxId:
      name: inboxId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: An error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    Slug:
      type: string
      pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
      minLength: 1
      maxLength: 64
      examples: [acme-app]
    LanguageTag:
      type: string
      description: A BCP 47 language tag.
      pattern: "^[a-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
      maxLength: 35
      examples: [en, tr]
    InboxMode:
      type: string
      description: |
        `live` shows who is available, typing and read receipts; `async` shows the expected reply
        time instead.
      enum: [live, async]
...
    BusinessHours:
      type: object
      description: When `enabled` is false the inbox counts as always open.
      required: [enabled, intervals]
      properties:
        enabled:
          type: boolean
        intervals:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/BusinessHoursInterval"
    InboxBranding:
      type: object
      properties:
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
          examples: ["#2563eb"]
        logo_url:
          type: string
          format: uri
          maxLength: 2000
        greeting:
          type: string
          maxLength: 500
    Inbox:
      type: object
      required: [id, name, slug, branding, default_locale, timezone, mode, business_hours, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          $ref: "#/components/schemas/Slug"
        branding:
          $ref: "#/components/schemas/InboxBranding"
        default_locale:
          $ref: "#/components/schemas/LanguageTag"
        timezone:
          type: string
          description: An IANA time zone name.
          examples: [Europe/Istanbul]
        mode:
          $ref: "#/components/schemas/InboxMode"
        expected_reply_minutes:
          type: integer
          format: int32
          description: Shown to contacts in `async` mode; absent when not set.
        business_hours:
          $ref: "#/components/schemas/BusinessHours"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    InboxUpdate:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          $ref: "#/components/schemas/Slug"
        branding:
          $ref: "#/components/schemas/InboxBranding"
        default_locale:
          $ref: "#/components/schemas/LanguageTag"
        timezone:
          type: string
          maxLength: 64
        mode:
          $ref: "#/components/schemas/InboxMode"
        expected_reply_minutes:
          type: [integer, "null"]
          format: int32
          minimum: 1
          maximum: 43200
          description: "`null` clears it."
        business_hours:
          $ref: "#/components/schemas/BusinessHours"
...
    ConversationEvent:
      type: object
      description: A conversation was created, or its status, snooze, assignee, priority, labels or subject changed.
      required: [id, type, workspace_id, inbox_id, conversation_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [conversation.created, conversation.updated]
        workspace_id:
          type: string
          format: uuid
        inbox_id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/Conversation"
    ConversationMovedEvent:
      type: object
      description: |
        A conversation left this event's inbox for another one. Sent to whoever can see the old
        inbox, so it can drop the conversation; `conversation.updated` follows for the new inbox.
      required: [id, type, workspace_id, inbox_id, conversation_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [conversation.moved]
        workspace_id:
          type: string
          format: uuid
        inbox_id:
          type: string
          format: uuid
          description: The inbox it left.
        conversation_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/ConversationMoved"
...
    MessageCreatedEvent:
      type: object
      description: A message, note or event message was added to a conversation.
      required: [id, type, workspace_id, inbox_id, conversation_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [message.created]
        workspace_id:
          type: string
          format: uuid
        inbox_id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/Message"
    DraftEvent:
      type: object
      description: |
        A draft was written, edited or discarded (`data` is the draft as it was). Sending a draft
        emits `message.created`.
      required: [id, type, workspace_id, inbox_id, conversation_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [draft.created, draft.updated, draft.deleted]
        workspace_id:
          type: string
          format: uuid
        inbox_id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/Message"
    MessageUpdatedEvent:
      type: object
      description: A message changed, e.g. its e-mail delivery state.
      required: [id, type, workspace_id, inbox_id, conversation_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [message.updated]
        workspace_id:
          type: string
          format: uuid
        inbox_id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/Message"
    ContactUpdatedEvent:
      type: object
      description: A contact changed.
      required: [id, type, workspace_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [contact.updated]
        workspace_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/Contact"
    ContactDeletedEvent:
      type: object
      description: A contact was deleted with its conversations.
      required: [id, type, workspace_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [contact.deleted]
        workspace_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/ContactRef"
...
    InboxCreatedEvent:
      type: object
      description: An inbox was created.
      required: [id, type, workspace_id, inbox_id, created_at, data]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [inbox.created]
        workspace_id:
          type: string
          format: uuid
        inbox_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/Inbox"
...
    RealtimeMessage:
      description: One server message on `/v1/realtime`, told apart by `type`.
      oneOf:
        - $ref: "#/components/schemas/ConversationEvent"
        - $ref: "#/components/schemas/ConversationMovedEvent"
        - $ref: "#/components/schemas/MessageCreatedEvent"
        - $ref: "#/components/schemas/MessageUpdatedEvent"
        - $ref: "#/components/schemas/DraftEvent"
        - $ref: "#/components/schemas/ContactUpdatedEvent"
        - $ref: "#/components/schemas/ContactDeletedEvent"
        - $ref: "#/components/schemas/InboxCreatedEvent"
        - $ref: "#/components/schemas/InboxUpdatedEvent"
        - $ref: "#/components/schemas/InboxDeletedEvent"
        - $ref: "#/components/schemas/InboxAccessChangedEvent"
        - $ref: "#/components/schemas/ConversationReadEvent"
        - $ref: "#/components/schemas/TypingEvent"
        - $ref: "#/components/schemas/RealtimeReady"
        - $ref: "#/components/schemas/RealtimeResyncRequired"
      discriminator:
        propertyName: type
        mapping:
          conversation.created: "#/components/schemas/ConversationEvent"
          conversation.updated: "#/components/schemas/ConversationEvent"
          conversation.moved: "#/components/schemas/ConversationMovedEvent"
          message.created: "#/components/schemas/MessageCreatedEvent"
          message.updated: "#/components/schemas/MessageUpdatedEvent"
          draft.created: "#/components/schemas/DraftEvent"
          draft.updated: "#/components/schemas/DraftEvent"
          draft.deleted: "#/components/schemas/DraftEvent"
          contact.updated: "#/components/schemas/ContactUpdatedEvent"
          contact.deleted: "#/components/schemas/ContactDeletedEvent"
          inbox.created: "#/components/schemas/InboxCreatedEvent"
          inbox.updated: "#/components/schemas/InboxUpdatedEvent"
          inbox.deleted: "#/components/schemas/InboxDeletedEvent"
          inbox_access.changed: "#/components/schemas/InboxAccessChangedEvent"
          conversation.read: "#/components/schemas/ConversationReadEvent"
          typing: "#/components/schemas/TypingEvent"
          ready: "#/components/schemas/RealtimeReady"
          resync_required: "#/components/schemas/RealtimeResyncRequired"
...
    Problem:
      type: object
      description: An RFC 9457 problem document.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          description: A URI that identifies the problem type.
        title:
          type: string
        status:
          type: integer
          format: int32
        code:
          type: string
          description: A stable, machine-readable error code.
          examples: [not_found]
        scope:
          type: string
          description: "`insufficient_scope` only: the scope the API key lacks."
          examples: [messages:write]
        detail:
          type: string